# docs/swagger.yaml
basePath: /api/v1
definitions:
  model.CreateDiscordBotRequest:
    properties:
      application_id:
        type: string
      client_id:
        type: string
      client_secret:
        type: string
      is_default:
        type: boolean
      name:
        maxLength: 100
        minLength: 2
        type: string
      public_key:
        type: string
      token:
        type: string
    required:
    - application_id
    - client_id
    - name
    - token
    type: object
  model.CreateDiscordChannelRequest:
    properties:
      bot_id:
        type: string
      channel_id:
        type: string
      description:
        type: string
      guild_id:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      webhook_url:
        type: string
    required:
    - bot_id
    - channel_id
    - name
    type: object
...
  model.DiscordBot:
    properties:
      application_id:
        type: string
      client_id:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_default:
        type: boolean
      name:
        type: string
      public_key:
        type: string
      updated_at:
        type: string
    type: object
  model.DiscordBotWithChannels:
    properties:
      application_id:
        type: string
      channels:
        items:
          $ref: '#/definitions/model.DiscordChannel'
        type: array
      client_id:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_default:
        type: boolean
      name:
        type: string
      public_key:
        type: string
      updated_at:
        type: string
    type: object
...
  model.UpdateDiscordBotRequest:
    properties:
      client_secret:
        type: string
      is_active:
        type: boolean
      is_default:
        type: boolean
      name:
        type: string
      token:
        type: string
    type: object
  model.UpdateDiscordChannelRequest:
    properties:
      description:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      webhook_url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Dynamic AI-Powered Task Scheduler with multi-provider AI integration
    for scraping jobs, news, and content, then intelligently processing and delivering
    notifications to Discord.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Multi-Worker API
  version: "1.0"
paths:
...
  /discord/bots:
    get:
      description: Get all registered Discord bots
      parameters:
      - description: Include inactive bots
        in: query
        name: include_inactive
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: List of bots
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: List Discord bots
      tags:
      - Discord Bots
    post:
      consumes:
      - application/json
      description: Register a new Discord bot with credentials
      parameters:
      - description: Bot credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateDiscordBotRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.DiscordBot'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Create a Discord bot
      tags:
      - Discord Bots
  /discord/bots/{botId}:
    delete:
      description: Delete a bot and all associated channels
      parameters:
      - description: Bot ID
        in: path
        name: botId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deletion status
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Bot not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Delete a Discord bot
      tags:
      - Discord Bots
    get:
      description: Get bot details with associated channels
      parameters:
      - description: Bot ID
        in: path
        name: botId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DiscordBotWithChannels'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Bot not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get a Discord bot
      tags:
      - Discord Bots
    put:
      consumes:
      - application/json
      description: Update bot settings or credentials
      parameters:
      - description: Bot ID
        in: path
        name: botId
        required: true
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateDiscordBotRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DiscordBot'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Update a Discord bot
      tags:
      - Discord Bots
  /discord/channels:
    get:
      description: Get all channels, optionally filtered by bot
      parameters:
      - description: Filter by bot ID
        in: query
        name: bot_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of channels
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: List Discord channels
      tags:
      - Discord Channels
    post:
      consumes:
      - application/json
      description: Register a channel with webhook for a bot
      parameters:
      - description: Channel configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateDiscordChannelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.DiscordChannel'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Create a Discord channel config
      tags:
      - Discord Channels
  /discord/channels/{channelId}:
    delete:
      description: Remove a channel configuration
      parameters:
      - description: Channel config ID
        in: path
        name: channelId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deletion status
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Channel not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Delete a Discord channel
      tags:
      - Discord Channels
    get:
      description: Get channel details
      parameters:
      - description: Channel config ID
        in: path
        name: channelId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DiscordChannel'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Channel not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get a Discord channel
      tags:
      - Discord Channels
    put:
      consumes:
      - application/json
      description: Update channel settings or webhook
      parameters:
      - description: Channel config ID
        in: path
        name: channelId
        required: true
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateDiscordChannelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DiscordChannel'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Update a Discord channel
      tags:
      - Discord Channels
  /discord/test:
    post:
      consumes:
      - application/json
      description: Send a test message to verify webhook configuration
      parameters:
      - description: Test webhook request
        in: body
        name: request
        required: true
        schema:
          properties:
            channel_id:
              type: string
            message:
              type: string
            webhook_url:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Test result
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Webhook error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Test Discord webhook
      tags:
      - Discord
...
  /tasks/{taskId}/discord:
    delete:
      description: Remove the Discord configuration for a task
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deletion status
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Delete task Discord config
      tags:
      - Task Discord Config
    get:
      description: Get the Discord configuration for a task
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TaskDiscordConfig'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Config not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get task Discord config
      tags:
      - Task Discord Config
    put:
      consumes:
      - application/json
      description: Configure which bot/channel/webhook a task uses for notifications
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: string
      - description: Discord configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.SetTaskDiscordConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TaskDiscordConfig'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Set task Discord config
      tags:
      - Task Discord Config
securityDefinitions:
  ApiKeyAuth:
    description: Enter your API key
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Enter your JWT token with the `Bearer ` prefix, e.g. "Bearer eyJhbGci..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"